Use `as_u32_or_panic ` to cast between `uszie` and `u32` (#2489)

+++ crates/module-system/sov-modules-api/src/state/accessors/genesis.rs
@@ -53,7 +53,7 @@ impl<'a, S: Spec> UniversalStateAccessor for GenesisStateAccessor<'a, S> {
         self.checkpoint.is_value_cached(namespace, key)
     }
 
-    fn get_size(&mut self, namespace: sov_state::Namespace, key: &SlotKey) -> Option<u64> {
+    fn get_size(&mut self, namespace: sov_state::Namespace, key: &SlotKey) -> Option<u32> {
         self.checkpoint.get_size(namespace, key)
     }
